1. удалила VacancyDtoFull 2. добавила в VacancyDto addressFull, areaId, areaName,areaParentId 3. mapper превратила в класс, тк понадобился Gson. Перенесла в data-слой 4. skills - список
