--- app/src/main/java/ru/practicum/android/diploma/data/dto/VacancyFullDto.kt
@@ -1,23 +1,0 @@
-package ru.practicum.android.diploma.data.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class VacancyFullDto(
-    val id: String,
-    val name: String,
-    val salary: SalaryDto?,
-    val address: AddressDto,
-    val experience: ExperienceDto?,
-    val schedule: ScheduleDto?,
-    val employment: EmploymentDto?,
-    val contacts: ContactsDto?,
-    val description: String,
-    val employer: EmployerDto?,
-    val area: AreaDto,
-    @SerializedName("response_url")
-    val responseUrl: Any,
-    @SerializedName("skills")
-    val skills: List<String>,
-    val url: String,
-    val industry: IndustryDto?
-)
