+++ app/src/main/java/ru/practicum/android/diploma/data/network/VacanciesAPI.kt
@@ -5,7 +5,7 @@ import retrofit2.http.Headers
 import retrofit2.http.Path
 import retrofit2.http.QueryMap
 import ru.practicum.android.diploma.BuildConfig
-import ru.practicum.android.diploma.data.dto.VacancyFullDto
+import ru.practicum.android.diploma.data.dto.VacancyDto
 
 interface VacanciesAPI {
 
@@ -23,5 +23,5 @@ interface VacanciesAPI {
     @GET("/vacancies/{vacancyId}")
     suspend fun getVacancyFull(
         @Path("vacancyId") vacancyId: String
-    ): VacancyFullDto
+    ): VacancyDto
 }
