+++ app/src/main/java/ru/practicum/android/diploma/data/network/VacanciesSearchResponse.kt
@@ -11,4 +11,3 @@ data class VacanciesSearchResponse(
     @SerializedName("per_page")
     val perPage: Int
 ) : Response()
-
